requirements.txt
##################################################################
# core python modules
##################################################################
Django==1.11.2
django_environ==0.4.3


##################################################################
# heroku hosting
##################################################################
whitenoise==3.3.0
psycopg2
django_redis
gunicorn==19.7.1

#
uwsgi
Flask

##################################################################
# django modules
##################################################################
django-admin-tools==0.8.1
django-braces==1.10.0
django-celery-beat==1.0.1
django-click==1.2.0
django-compressor==2.0
django-split-settings==0.2.5

##################################################################
# RESTful API
# http://www.django-rest-framework.org/#installation
##################################################################
djangorestframework==3.6.3
markdown==2.6.8
coreapi==2.3.1

##################################################################
# authentication
##################################################################
django-authtools==1.5.0

##################################################################
# task handling
##################################################################
celery[redis]==4.0.2

##################################################################
# other modules
##################################################################
raven==6.0.0


##################################################################
# external modules & patches
##################################################################
-e git+https://github.com/palmbeach-interactive/django-slick-admin@master#egg=django-slick-admin
-e git+https://github.com/spotify/echoprint-server@master#egg=echoprint-server
